Add additional tests
