+++ src/components/Wizards/RosaWizard/Steps/BasicSetupStep/DetailsSubStep/DetailsSubStep.spec.tsx
@@ -1,8 +1,9 @@
 import { test, expect } from '@playwright/experimental-ct-react';
 import { checkAccessibility } from '../../../../../../test-helpers';
-import { DetailsSubStepStory } from './DetailsSubStep.story';
-import { mockSingleBillingAccount, mockRegions } from './DetailsSubStep.story';
-import type { Resource, SelectDropdownType } from '../../../../types';
+import type { OpenShiftVersionsData, Resource, Role, SelectDropdownType } from '../../../../types';
+import { defaultRosaWizardStrings } from '../../../rosaWizardStrings.defaults';
+import { DetailsSubStepMount } from './DetailsSubStep.spec-helpers';
+import { mockRegions, mockSingleBillingAccount } from './DetailsSubStep.fixtures';
 
 const mockResource = <TData,>(data: TData): Resource<TData> => ({
   data,
@@ -11,21 +12,65 @@ const mockResource = <TData,>(data: TData): Resource<TData> => ({
   fetch: async () => {},
 });
 
+/** latest, default, and previous releases differ so grouping yields three labeled sections. */
+const mockVersionsLatestDefaultPrevious: OpenShiftVersionsData = {
+  latest: { label: 'OpenShift 4.14.0', value: '4.14.0' },
+  default: { label: 'OpenShift 4.13.1', value: '4.13.1' },
+  releases: [
+    { label: 'OpenShift 4.12.0', value: '4.12.0' },
+    { label: 'OpenShift 4.11.5', value: '4.11.5' },
+  ],
+};
+
+const mockVersionsDefaultEqualsLatest: OpenShiftVersionsData = {
+  latest: { label: 'OpenShift 4.12.0', value: '4.12.0' },
+  default: { label: 'OpenShift 4.12.0', value: '4.12.0' },
+  releases: [{ label: 'OpenShift 4.11.5', value: '4.11.5' }],
+};
+
+const INSTALLER_ARN = 'arn:aws:iam::123456789012:role/rosa-installer';
+
+const rolesWithInstallerVersion412: Resource<Role[], [awsAccount: string]> & {
+  fetch: (awsAccount: string) => Promise<void>;
+} = {
+  data: [
+    {
+      installerRole: {
+        label: 'Installer role A',
+        value: INSTALLER_ARN,
+        roleVersion: '4.12.0',
+      },
+      supportRole: [{ label: 'Support Role', value: 'arn:aws:iam::123456789012:role/support' }],
+      workerRole: [{ label: 'Worker Role', value: 'arn:aws:iam::123456789012:role/worker' }],
+    },
+  ],
+  isFetching: false,
+  error: null,
+  fetch: async () => {},
+};
+
+const versionDisabledDescription =
+  defaultRosaWizardStrings.details.openShiftVersionOptionDisabledDescription;
+
 test.describe('DetailsSubStep', () => {
   test('should pass accessibility tests', async ({ mount }) => {
-    const component = await mount(<DetailsSubStepStory />);
+    const component = await mount(<DetailsSubStepMount />);
     await checkAccessibility({ component });
   });
 
-  test('should render OpenShift version select', async ({ mount }) => {
-    const component = await mount(<DetailsSubStepStory />);
+  test('should render the Details section title', async ({ mount }) => {
+    const component = await mount(<DetailsSubStepMount />);
+    await expect(component.getByText('Details', { exact: true })).toBeVisible();
+  });
 
+  test('should render OpenShift version select', async ({ mount }) => {
+    const component = await mount(<DetailsSubStepMount />);
     await expect(component.getByText('OpenShift version', { exact: true })).toBeVisible();
   });
 
   test('should render with empty options', async ({ mount }) => {
     const component = await mount(
-      <DetailsSubStepStory
+      <DetailsSubStepMount
         versions={{
           data: { releases: [] },
           error: null,
@@ -43,14 +88,14 @@ test.describe('DetailsSubStep', () => {
   });
 
   test('should render the Cluster name input', async ({ mount }) => {
-    const component = await mount(<DetailsSubStepStory />);
+    const component = await mount(<DetailsSubStepMount />);
 
     await expect(component.getByText('Cluster name', { exact: true })).toBeVisible();
     await expect(component.getByRole('textbox', { name: 'Cluster name' })).toBeVisible();
   });
 
-  test('should render AWS infrastructure account select', async ({ mount }) => {
-    const component = await mount(<DetailsSubStepStory />);
+  test('should render the Associated AWS infrastructure account select', async ({ mount }) => {
+    const component = await mount(<DetailsSubStepMount />);
 
     await expect(
       component.getByText('Associated AWS infrastructure account', { exact: true })
@@ -58,7 +103,7 @@ test.describe('DetailsSubStep', () => {
   });
 
   test('should render the Associated AWS billing account select', async ({ mount }) => {
-    const component = await mount(<DetailsSubStepStory />);
+    const component = await mount(<DetailsSubStepMount />);
 
     await expect(
       component.getByText('Associated AWS billing account', { exact: true })
@@ -66,19 +111,19 @@ test.describe('DetailsSubStep', () => {
   });
 
   test('should render the Region select', async ({ mount }) => {
-    const component = await mount(<DetailsSubStepStory />);
+    const component = await mount(<DetailsSubStepMount />);
 
     await expect(component.getByText('Region', { exact: true })).toBeVisible();
   });
 
   test('should render the Associate a new AWS account button', async ({ mount }) => {
-    const component = await mount(<DetailsSubStepStory />);
+    const component = await mount(<DetailsSubStepMount />);
 
     await expect(component.getByText('Associate a new AWS account')).toBeVisible();
   });
 
   test('should show OpenShift version options in dropdown', async ({ mount, page }) => {
-    const component = await mount(<DetailsSubStepStory />);
+    const component = await mount(<DetailsSubStepMount />);
 
     const versionCombobox = component.locator('#cluster-cluster_version [role="combobox"]');
     await versionCombobox.click();
@@ -91,7 +136,7 @@ test.describe('DetailsSubStep', () => {
     test('should display validation error for cluster name with uppercase characters', async ({
       mount,
     }) => {
-      const component = await mount(<DetailsSubStepStory />);
+      const component = await mount(<DetailsSubStepMount />);
 
       const nameInput = component.getByRole('textbox', { name: /Cluster name/ });
       await nameInput.fill('MyCluster');
@@ -103,7 +148,7 @@ test.describe('DetailsSubStep', () => {
     test('should display validation error for cluster name starting with a number', async ({
       mount,
     }) => {
-      const component = await mount(<DetailsSubStepStory />);
+      const component = await mount(<DetailsSubStepMount />);
 
       const nameInput = component.getByRole('textbox', { name: /Cluster name/ });
       await nameInput.fill('1cluster');
@@ -115,7 +160,7 @@ test.describe('DetailsSubStep', () => {
     test('should cancel pending async check when name becomes sync-invalid', async ({ mount }) => {
       const calls: Array<{ name: string; region: string }> = [];
       const component = await mount(
-        <DetailsSubStepStory
+        <DetailsSubStepMount
           checkClusterNameUniqueness={(name, region) => {
             if (region) calls.push({ name, region });
           }}
@@ -124,15 +169,13 @@ test.describe('DetailsSubStep', () => {
       );
       const nameInput = component.getByRole('textbox', { name: /Cluster name/ });
       await nameInput.fill('taken');
-      // Immediately change to invalid before debounce fires
       await nameInput.fill('TAKEN');
       await component.page().waitForTimeout(700);
-      // The old "taken" check should never have fired
       expect(calls.length).toBe(0);
     });
 
     test('should show AWS infrastructure account options in dropdown', async ({ mount, page }) => {
-      const component = await mount(<DetailsSubStepStory />);
+      const component = await mount(<DetailsSubStepMount />);
 
       const awsCombobox = component.locator('#cluster-associated_aws_id [role="combobox"]');
       await awsCombobox.click();
@@ -146,7 +189,7 @@ test.describe('DetailsSubStep', () => {
     });
 
     test('should show region options in dropdown', async ({ mount, page }) => {
-      const component = await mount(<DetailsSubStepStory />);
+      const component = await mount(<DetailsSubStepMount />);
 
       const regionCombobox = component.locator('#cluster-region [role="combobox"]');
       await regionCombobox.click();
@@ -160,7 +203,7 @@ test.describe('DetailsSubStep', () => {
       mount,
     }) => {
       const component = await mount(
-        <DetailsSubStepStory
+        <DetailsSubStepMount
           awsInfrastructureAccounts={{ data: [], isFetching: true, error: null }}
         />
       );
@@ -172,7 +215,7 @@ test.describe('DetailsSubStep', () => {
 
     test('should show disabled state for AWS billing account when loading', async ({ mount }) => {
       const component = await mount(
-        <DetailsSubStepStory awsBillingAccounts={{ data: [], isFetching: true, error: null }} />
+        <DetailsSubStepMount awsBillingAccounts={{ data: [], isFetching: true, error: null }} />
       );
 
       const billingSelect = component.locator('#cluster-billing_account_id');
@@ -182,7 +225,7 @@ test.describe('DetailsSubStep', () => {
 
     test('should show disabled state for Region select when loading', async ({ mount }) => {
       const component = await mount(
-        <DetailsSubStepStory
+        <DetailsSubStepMount
           regions={{ data: [], isFetching: true, error: null, fetch: async () => {} }}
         />
       );
@@ -194,7 +237,7 @@ test.describe('DetailsSubStep', () => {
 
     test('should auto-select billing account when only one is available', async ({ mount }) => {
       const component = await mount(
-        <DetailsSubStepStory
+        <DetailsSubStepMount
           awsBillingAccounts={{ data: mockSingleBillingAccount, isFetching: false, error: null }}
         />
       );
@@ -205,7 +248,7 @@ test.describe('DetailsSubStep', () => {
 
     test('should not display async error when sync validation fails', async ({ mount }) => {
       const component = await mount(
-        <DetailsSubStepStory
+        <DetailsSubStepMount
           clusterNameValidation={{
             error: 'Cluster name already exists.',
             isFetching: false,
@@ -225,7 +268,7 @@ test.describe('DetailsSubStep', () => {
     test('should not auto-select billing account when multiple are available', async ({
       mount,
     }) => {
-      const component = await mount(<DetailsSubStepStory />);
+      const component = await mount(<DetailsSubStepMount />);
 
       const billingCombobox = component.locator('#cluster-billing_account_id [role="combobox"]');
       await expect(billingCombobox).toHaveValue('');
@@ -233,7 +276,7 @@ test.describe('DetailsSubStep', () => {
 
     test('should render with empty OpenShift versions', async ({ mount }) => {
       const component = await mount(
-        <DetailsSubStepStory
+        <DetailsSubStepMount
           versions={{
             data: { releases: [] },
             isFetching: false,
@@ -248,7 +291,7 @@ test.describe('DetailsSubStep', () => {
 
     test('should render with empty regions', async ({ mount }) => {
       const component = await mount(
-        <DetailsSubStepStory
+        <DetailsSubStepMount
           regions={{ data: [], isFetching: false, error: null, fetch: async () => {} }}
         />
       );
@@ -256,14 +299,16 @@ test.describe('DetailsSubStep', () => {
       await expect(component.getByText('Region', { exact: true })).toBeVisible();
     });
 
-    test('should render the Connect ROSA to a new AWS billing account link', async ({ mount }) => {
-      const component = await mount(<DetailsSubStepStory />);
+    test('should render the Connect ROSA to a new AWS billing account link', async ({
+      mount,
+    }) => {
+      const component = await mount(<DetailsSubStepMount />);
 
       await expect(component.getByText('Connect ROSA to a new AWS billing account')).toBeVisible();
     });
 
     test('should allow typing a cluster name', async ({ mount }) => {
-      const component = await mount(<DetailsSubStepStory />);
+      const component = await mount(<DetailsSubStepMount />);
 
       const nameInput = component.getByRole('textbox', { name: 'Cluster name' });
       await nameInput.fill('my-test-cluster');
@@ -272,7 +317,7 @@ test.describe('DetailsSubStep', () => {
     });
 
     test('should select an OpenShift version', async ({ mount, page }) => {
-      const component = await mount(<DetailsSubStepStory />);
+      const component = await mount(<DetailsSubStepMount />);
 
       const versionCombobox = component.locator('#cluster-cluster_version [role="combobox"]');
       await versionCombobox.click();
@@ -282,7 +327,7 @@ test.describe('DetailsSubStep', () => {
     });
 
     test('should select a region', async ({ mount, page }) => {
-      const component = await mount(<DetailsSubStepStory />);
+      const component = await mount(<DetailsSubStepMount />);
 
       const regionCombobox = component.locator('#cluster-region [role="combobox"]');
       await regionCombobox.click();
@@ -293,7 +338,7 @@ test.describe('DetailsSubStep', () => {
 
     test('should render pre-filled cluster data', async ({ mount }) => {
       const component = await mount(
-        <DetailsSubStepStory
+        <DetailsSubStepMount
           clusterOverrides={{
             name: 'existing-cluster',
             billing_account_id: 'billing-main-123456789012',
@@ -313,7 +358,7 @@ test.describe('DetailsSubStep', () => {
       const calls: Array<{ name: string; region: string }> = [];
 
       const component = await mount(
-        <DetailsSubStepStory
+        <DetailsSubStepMount
           checkClusterNameUniqueness={(name, region) => {
             if (region) calls.push({ name, region });
           }}
@@ -324,7 +369,6 @@ test.describe('DetailsSubStep', () => {
       const nameInput = component.getByRole('textbox', { name: /Cluster name/ });
       await nameInput.fill('valid-cluster');
 
-      // wait for debounce (500ms) + buffer
       await component.page().waitForTimeout(700);
 
       expect(calls.length).toBeGreaterThanOrEqual(1);
@@ -335,7 +379,7 @@ test.describe('DetailsSubStep', () => {
       const calls: Array<{ name: string; region: string }> = [];
 
       const component = await mount(
-        <DetailsSubStepStory
+        <DetailsSubStepMount
           checkClusterNameUniqueness={(name, region) => {
             if (region) calls.push({ name, region });
           }}
@@ -357,7 +401,7 @@ test.describe('DetailsSubStep', () => {
       const calls: Array<{ name: string; region: string }> = [];
 
       const component = await mount(
-        <DetailsSubStepStory
+        <DetailsSubStepMount
           checkClusterNameUniqueness={(name, region) => {
             if (region) calls.push({ name, region });
           }}
@@ -383,7 +427,7 @@ test.describe('DetailsSubStep', () => {
       const calls: Array<{ name: string; region: string }> = [];
 
       const component = await mount(
-        <DetailsSubStepStory
+        <DetailsSubStepMount
           checkClusterNameUniqueness={(name, region) => {
             if (region) calls.push({ name, region });
           }}
@@ -409,7 +453,7 @@ test.describe('DetailsSubStep', () => {
       const calls: Array<{ name: string; region?: string }> = [];
 
       const component = await mount(
-        <DetailsSubStepStory
+        <DetailsSubStepMount
           checkClusterNameUniqueness={(name, region) => {
             calls.push({ name, region });
           }}
@@ -429,7 +473,7 @@ test.describe('DetailsSubStep', () => {
       mount,
     }) => {
       const component = await mount(
-        <DetailsSubStepStory clusterOverrides={{ region: 'us-east-1' }} />
+        <DetailsSubStepMount clusterOverrides={{ region: 'us-east-1' }} />
       );
 
       const nameInput = component.getByRole('textbox', { name: /Cluster name/ });
@@ -439,5 +483,129 @@ test.describe('DetailsSubStep', () => {
 
       await expect(component.getByText('Details', { exact: true })).toBeVisible();
     });
+  });
+
+  test('version dropdown displays latest, default, and other versions correctly in grouped sections', async ({
+    mount,
+    page,
+  }) => {
+    const component = await mount(
+      <DetailsSubStepMount
+        versions={{
+          data: mockVersionsLatestDefaultPrevious,
+          isFetching: false,
+          error: null,
+          fetch: async () => {},
+        }}
+      />
+    );
+
+    const versionCombobox = component.getByRole('combobox', {
+      name: 'Select an OpenShift version',
+    });
+    await versionCombobox.click();
+
+    const latestGroup = page.locator('section').filter({
+      has: page.getByRole('heading', { name: 'Latest release', exact: true }),
+    });
+    await expect(latestGroup.getByText('OpenShift 4.14.0', { exact: true })).toBeVisible();
+
+    const defaultGroup = page.locator('section').filter({
+      has: page.getByRole('heading', { name: 'Default release', exact: true }),
+    });
+    await expect(defaultGroup.getByText('OpenShift 4.13.1', { exact: true })).toBeVisible();
+
+    const previousGroup = page.locator('section').filter({
+      has: page.getByRole('heading', { name: 'Previous releases', exact: true }),
+    });
+    await expect(previousGroup.getByText('OpenShift 4.12.0', { exact: true })).toBeVisible();
+    await expect(previousGroup.getByText('OpenShift 4.11.5', { exact: true })).toBeVisible();
+  });
+
+  test('version dropdown shows Default (Recommended) and Previous releases only when latest and default share the same value', async ({
+    mount,
+    page,
+  }) => {
+    const component = await mount(
+      <DetailsSubStepMount
+        versions={{
+          data: mockVersionsDefaultEqualsLatest,
+          isFetching: false,
+          error: null,
+          fetch: async () => {},
+        }}
+      />
+    );
+
+    const versionCombobox = component.getByRole('combobox', {
+      name: 'Select an OpenShift version',
+    });
+    await versionCombobox.click();
+
+    await expect(
+      page.getByRole('heading', { name: 'Default (Recommended)', exact: true })
+    ).toBeVisible();
+    await expect(
+      page.getByRole('heading', { name: 'Previous releases', exact: true })
+    ).toBeVisible();
+    await expect(
+      page.getByRole('heading', { name: 'Latest release', exact: true })
+    ).not.toBeVisible();
+    await expect(
+      page.getByRole('heading', { name: 'Default release', exact: true })
+    ).not.toBeVisible();
+
+    const recommendedGroup = page.locator('section').filter({
+      has: page.getByRole('heading', { name: 'Default (Recommended)', exact: true }),
+    });
+    await expect(recommendedGroup.getByText('OpenShift 4.12.0', { exact: true })).toBeVisible();
+  });
+
+  test('should disable OpenShift version options newer than the selected installer role version', async ({
+    mount,
+    page,
+  }) => {
+    const component = await mount(
+      <DetailsSubStepMount
+        roles={rolesWithInstallerVersion412}
+        clusterOverrides={{ installer_role_arn: INSTALLER_ARN }}
+        versions={{
+          data: mockVersionsLatestDefaultPrevious,
+          isFetching: false,
+          error: null,
+          fetch: async () => {},
+        }}
+      />
+    );
+
+    await component.getByRole('combobox', { name: 'Select an OpenShift version' }).click();
+
+    await expect(page.getByRole('option', { name: /OpenShift 4\.14\.0/ })).toBeDisabled();
+    await expect(page.getByRole('option', { name: /OpenShift 4\.13\.1/ })).toBeDisabled();
+    await expect(page.getByRole('option', { name: /^OpenShift 4\.12\.0$/ })).toBeEnabled();
+    await expect(page.getByRole('option', { name: /^OpenShift 4\.11\.5$/ })).toBeEnabled();
+  });
+
+  test('should show description on disabled OpenShift version options when installer role is older', async ({
+    mount,
+    page,
+  }) => {
+    const component = await mount(
+      <DetailsSubStepMount
+        roles={rolesWithInstallerVersion412}
+        clusterOverrides={{ installer_role_arn: INSTALLER_ARN }}
+        versions={{
+          data: mockVersionsLatestDefaultPrevious,
+          isFetching: false,
+          error: null,
+          fetch: async () => {},
+        }}
+      />
+    );
+
+    await component.getByRole('combobox', { name: 'Select an OpenShift version' }).click();
+
+    await expect(page.getByText(versionDisabledDescription, { exact: true }).first()).toBeVisible();
+    await expect(page.getByText(versionDisabledDescription, { exact: true })).toHaveCount(2);
   });
 });
